Split profile popup component
So that it can be used both in the bottom tray and the floating popup (they are not in the same module, and one of them doesn't require the popup wrapper around)
