+++ src/modules/drawing/components/DrawingTooltip.vue
