+++ src/modules/drawing/components/ProfilePopup.vue
@@ -1,12 +1,7 @@
 <template>
     <div v-if="showProfile" data-cy="profile-popup" class="profile-popup">
         <div ref="profile-graph" class="profile-graph"></div>
-        <div
-            v-show="showTooltip"
-            ref="profile-tooltip"
-            class="profile-tooltip"
-            data-cy="profile-popup-tooltip"
-        >
+        <div ref="profile-tooltip" class="profile-tooltip" data-cy="profile-popup-tooltip">
             <div class="profile-tooltip-inner p-2">
                 <div>
                     <strong>Distance: </strong>
@@ -19,56 +14,17 @@
             </div>
             <div class="profile-tooltip-arrow"></div>
         </div>
-        <div v-if="profileInformation.length" class="d-flex">
-            <div
-                class="flex-grow-1 profile-info-container d-flex"
-                data-cy="profile-popup-info-container"
-            >
-                <span
-                    v-for="info in profileInformation"
-                    :key="info.title"
-                    :title="$t(info.title)"
-                    class="mx-2 text-nowrap d-flex align-content-center align-self-center"
-                >
-                    <font-awesome-icon
-                        v-for="(icon, index) in info.icons"
-                        :key="`${info.title}-${index}`"
-                        :icon="icon"
-                        class="me-1 align-self-center"
-                    />
-                    &nbsp;
-                    <span data-cy="profile-popup-info">
-                        {{ info.value }}
-                    </span>
-                </span>
-            </div>
-            <ButtonWithIcon
-                :button-font-awesome-icon="['far', 'trash-alt']"
-                data-cy="profile-popup-delete-button"
-                @click="onDelete"
-            />
-        </div>
+        <ProfilePopupPlot class="card-body p-2" :feature="feature" @delete="onDelete" />
     </div>
 </template>
 
 <script>
 import { EditableFeature, EditableFeatureTypes } from '@/api/features.api'
-import { profile } from '@/api/profile.api'
-import { formatTime, toLv95 } from '@/modules/drawing/lib/drawingUtils'
-import ProfileChart from '@/modules/drawing/lib/ProfileChart'
-import { sketchPointStyle } from '@/modules/drawing/lib/style'
+import ProfilePopupPlot from '@/modules/drawing/components/ProfilePopupPlot.vue'
 import { UIModes } from '@/modules/store/modules/ui.store'
-import ButtonWithIcon from '@/utils/ButtonWithIcon.vue'
-import { format } from '@/utils/numberUtils'
-import * as d3 from 'd3'
-import { Feature } from 'ol'
-import { Point } from 'ol/geom'
-import { Vector as VectorLayer } from 'ol/layer'
-import { Vector as VectorSource } from 'ol/source'
-import proj4 from 'proj4'
 
 export default {
-    components: { ButtonWithIcon },
+    components: { ProfilePopupPlot },
     inject: ['getMap'],
     props: {
         feature: {
@@ -81,19 +37,6 @@ export default {
         },
     },
     emits: ['close', 'delete'],
-    data() {
-        return {
-            options: {
-                margin: { left: 50, right: 0, bottom: 35, top: 5 },
-                width: 0,
-                height: 0,
-                xLabel: 'profile_x_label',
-                yLabel: 'profile_y_label',
-            },
-            showTooltip: false,
-            profileInfo: null,
-        }
-    },
     computed: {
         showProfile() {
             return (
@@ -103,95 +46,6 @@ export default {
                 )
             )
         },
-        profileInformation() {
-            if (!this.profileInfo) {
-                return []
-            }
-            return [
-                {
-                    title: 'profile_elevation_difference',
-                    icons: [['fa', 'arrows-alt-v']],
-                    value: this.formatElevation(this.profileInfo.elevationDiff),
-                },
-                {
-                    title: 'profile_elevation_up',
-                    icons: [['fa', 'sort-amount-up-alt']],
-                    value: this.formatElevation(this.profileInfo.totalElevationDiff[0]),
-                },
-                {
-                    title: 'profile_elevation_down',
-                    icons: [['fa', 'sort-amount-down-alt']],
-                    value: this.formatElevation(this.profileInfo.totalElevationDiff[1]),
-                },
-                {
-                    title: 'profile_poi_up',
-                    icons: [['fa', 'chevron-up']],
-                    value: this.formatElevation(this.profileInfo.elevationPoints[0]),
-                },
-                {
-                    title: 'profile_poi_down',
-                    icons: [['fa', 'chevron-down']],
-                    value: this.formatElevation(this.profileInfo.elevationPoints[1]),
-                },
-                {
-                    title: 'profile_distance',
-                    icons: [
-                        ['far', 'eye'],
-                        ['fa', 'arrows-alt-h'],
-                    ],
-                    value: this.formatDistance(this.profileInfo.distance),
-                },
-                {
-                    title: 'profile_slope_distance',
-                    icons: [['fa', 'arrows-alt-h']],
-                    value: this.formatDistance(this.profileInfo.slopeDistance),
-                },
-                {
-                    title: 'profile_hike_time',
-                    icons: [['far', 'clock']],
-                    value: formatTime(this.profileInfo.hikingTime),
-                },
-            ]
-        },
-    },
-    watch: {
-        feature: {
-            deep: true,
-            handler() {
-                this.triggerProfileUpdate = true
-            },
-        },
-    },
-    created() {
-        this.positionOnMap = new Point([0, 0])
-        /** Additional overlay to display azimuth circle */
-        this.overlay = new VectorLayer({
-            source: new VectorSource({
-                useSpatialIndex: false,
-                features: [new Feature(this.positionOnMap)],
-            }),
-            style: sketchPointStyle,
-            updateWhileAnimating: true,
-            updateWhileInteracting: true,
-            zIndex: 2000,
-        })
-    },
-    async mounted() {
-        // listening to window.resize event so that we resize the SVG profile
-        window.addEventListener('resize', this.onResize)
-        await this.updateProfile()
-    },
-    unmounted() {
-        window.removeEventListener('resize', this.onResize)
-    },
-    updated() {
-        this.$nextTick(function () {
-            // Code that will run only after the entire view has been re-rendered
-            if (this.$refs['profile-graph'] && this.triggerProfileUpdate) {
-                this.triggerProfileUpdate = false
-                this.updateProfile()
-            }
-        })
     },
     methods: {
         onClose() {
@@ -199,113 +53,6 @@ export default {
         },
         onDelete() {
             this.$emit('delete')
-        },
-        onResize() {
-            this.triggerProfileUpdate = true
-            this.updateProfile()
-        },
-        async updateProfile() {
-            const containerEl = this.$refs['profile-graph']
-            if (!containerEl) {
-                return
-            }
-            let chartEl
-            if (!this.profileChart) {
-                this.options.width = containerEl.clientWidth
-                this.options.height = containerEl.clientHeight
-                this.profileChart = new ProfileChart(this.options)
-                chartEl = await this.createProfileChart()
-            } else {
-                chartEl = await this.updateProfileChart([
-                    containerEl.clientWidth,
-                    containerEl.clientHeight,
-                ])
-            }
-            if (chartEl) {
-                this.profileInfo = this.profileChart.getProfileInfo()
-                containerEl.appendChild(chartEl)
-            }
-        },
-        async createProfileChart() {
-            const data = await this.getProfile(this.feature.coordinates)
-            this.profileChart.create(data)
-            const areaChartPath = this.profileChart.group.select('.profile-area')
-            this.attachPathListeners(areaChartPath)
-            return this.profileChart.element
-        },
-        async updateProfileChart(size) {
-            const data = await this.getProfile(this.feature.coordinates)
-            this.profileChart.update(data, size)
-            return this.profileChart.element
-        },
-        async getProfile(coordinates) {
-            const coordinatesLv95 = toLv95(coordinates, 'EPSG:3857')
-            return await profile({
-                geom: `{"type":"LineString","coordinates":${JSON.stringify(coordinatesLv95)}}`,
-                offset: 0,
-                sr: 2056,
-                distinct_points: true,
-            })
-        },
-        attachPathListeners(areaChartPath) {
-            areaChartPath.on('mousemove', (evt) => {
-                const x = d3.pointer(evt)[0]
-                let pos = evt.target.getPointAtLength(x)
-
-                const start = x
-                const end = pos.x
-                const accuracy = 5
-                for (let i = start; i > end; i += accuracy) {
-                    pos = evt.target.getPointAtLength(i)
-                    if (pos.x >= x) {
-                        break
-                    }
-                }
-
-                // Get the coordinate value of x and y
-                const xCoord = this.profileChart.domain.X.invert(x)
-                const yCoord = this.profileChart.domain.Y.invert(pos.y)
-                const positionX = this.profileChart.domain.X(xCoord) + this.options.margin.left
-                const positionY = this.profileChart.domain.Y(yCoord) + this.options.margin.top
-                const toltipEl = this.$refs['profile-tooltip']
-                // done like this because using of computed makes it very slow
-                toltipEl.style.left = `${positionX}px`
-                toltipEl.style.top = `${positionY}px`
-                toltipEl.querySelector('.distance').innerText = `${xCoord.toFixed(2)}${
-                    this.profileInfo.unitX
-                }`
-                toltipEl.querySelector('.elevation').innerText = `${yCoord.toFixed(2)} m`
-                const coordsMap = this.profileChart.findMapCoordinates(xCoord)
-                this.positionOnMap.setCoordinates(proj4('EPSG:2056', 'EPSG:3857', coordsMap))
-            })
-
-            areaChartPath.on('mouseover', () => {
-                this.showTooltip = true
-                this.overlay.setMap(this.getMap())
-            })
-
-            areaChartPath.on('mouseout', () => {
-                this.showTooltip = false
-                this.overlay.setMap(null)
-            })
-        },
-        formatDistance(value) {
-            if (isNaN(value) || value === null) {
-                return '0.00m'
-            }
-            if (value < 1000) {
-                return `${value.toFixed(2)}m`
-            }
-            return `${(value / 1000).toFixed(2)}km`
-        },
-        formatElevation(value) {
-            if (isNaN(value) || value === null) {
-                return '0.00m'
-            }
-            if (value < 1000) {
-                return `${value.toFixed(2)}m`
-            }
-            return `${format(Math.round(value), 3)}m`
         },
     },
 }
@@ -321,57 +68,6 @@ export default {
     .profile-info-container {
         overflow-x: auto;
         max-width: 100vw;
-    }
-    .profile-graph {
-        overflow: hidden;
-        width: 100%;
-        height: 145px;
-
-        .axis path,
-        .axis line {
-            fill: none;
-            stroke: #000;
-            shape-rendering: crispEdges;
-        }
-        .tick {
-            font-size: 11px;
-        }
-        .profile-inner {
-            overflow: hidden;
-            width: 100%;
-            height: 145px;
-
-            .profile-svg {
-                display: block;
-                margin: 0;
-                width: 100%;
-                height: 100%;
-            }
-        }
-        .profile-grid-x,
-        .profile-grid-y {
-            stroke: black;
-            opacity: 0.8;
-
-            line {
-                stroke-width: 0.02em;
-            }
-        }
-        .profile-area {
-            fill: $primary;
-            fill-opacity: 0.5;
-        }
-        .profile-legend,
-        .profile-label {
-            font-weight: bold;
-            text-shadow: 1px 1px $light;
-        }
-        svg {
-            overflow: visible;
-        }
-        text {
-            cursor: default;
-        }
     }
 
     .profile-tooltip {
